+++ db/structure.sql
@@ -367,6 +367,72 @@ ALTER SEQUENCE public.entities_id_seq OWNED BY public.entities.id;
 
 
 --
+-- Name: event_store_events; Type: TABLE; Schema: public; Owner: -
+--
+
+CREATE TABLE public.event_store_events (
+    id bigint NOT NULL,
+    event_id uuid NOT NULL,
+    event_type character varying NOT NULL,
+    metadata jsonb,
+    data jsonb NOT NULL,
+    created_at timestamp(6) without time zone NOT NULL,
+    valid_at timestamp(6) without time zone
+);
+
+
+--
+-- Name: event_store_events_id_seq; Type: SEQUENCE; Schema: public; Owner: -
+--
+
+CREATE SEQUENCE public.event_store_events_id_seq
+    START WITH 1
+    INCREMENT BY 1
+    NO MINVALUE
+    NO MAXVALUE
+    CACHE 1;
+
+
+--
+-- Name: event_store_events_id_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: -
+--
+
+ALTER SEQUENCE public.event_store_events_id_seq OWNED BY public.event_store_events.id;
+
+
+--
+-- Name: event_store_events_in_streams; Type: TABLE; Schema: public; Owner: -
+--
+
+CREATE TABLE public.event_store_events_in_streams (
+    id bigint NOT NULL,
+    stream character varying NOT NULL,
+    "position" integer,
+    event_id uuid NOT NULL,
+    created_at timestamp(6) without time zone NOT NULL
+);
+
+
+--
+-- Name: event_store_events_in_streams_id_seq; Type: SEQUENCE; Schema: public; Owner: -
+--
+
+CREATE SEQUENCE public.event_store_events_in_streams_id_seq
+    START WITH 1
+    INCREMENT BY 1
+    NO MINVALUE
+    NO MAXVALUE
+    CACHE 1;
+
+
+--
+-- Name: event_store_events_in_streams_id_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: -
+--
+
+ALTER SEQUENCE public.event_store_events_in_streams_id_seq OWNED BY public.event_store_events_in_streams.id;
+
+
+--
 -- Name: identities; Type: TABLE; Schema: public; Owner: -
 --
 
@@ -907,6 +973,20 @@ ALTER TABLE ONLY public.entities ALTER COLUMN id SET DEFAULT nextval('public.ent
 
 
 --
+-- Name: event_store_events id; Type: DEFAULT; Schema: public; Owner: -
+--
+
+ALTER TABLE ONLY public.event_store_events ALTER COLUMN id SET DEFAULT nextval('public.event_store_events_id_seq'::regclass);
+
+
+--
+-- Name: event_store_events_in_streams id; Type: DEFAULT; Schema: public; Owner: -
+--
+
+ALTER TABLE ONLY public.event_store_events_in_streams ALTER COLUMN id SET DEFAULT nextval('public.event_store_events_in_streams_id_seq'::regclass);
+
+
+--
 -- Name: identities id; Type: DEFAULT; Schema: public; Owner: -
 --
 
@@ -1034,6 +1114,22 @@ ALTER TABLE ONLY public.companies
 
 ALTER TABLE ONLY public.entities
     ADD CONSTRAINT entities_pkey PRIMARY KEY (id);
+
+
+--
+-- Name: event_store_events_in_streams event_store_events_in_streams_pkey; Type: CONSTRAINT; Schema: public; Owner: -
+--
+
+ALTER TABLE ONLY public.event_store_events_in_streams
+    ADD CONSTRAINT event_store_events_in_streams_pkey PRIMARY KEY (id);
+
+
+--
+-- Name: event_store_events event_store_events_pkey; Type: CONSTRAINT; Schema: public; Owner: -
+--
+
+ALTER TABLE ONLY public.event_store_events
+    ADD CONSTRAINT event_store_events_pkey PRIMARY KEY (id);
 
 
 --
@@ -1219,6 +1315,55 @@ CREATE UNIQUE INDEX index_entities_on_provider_and_external_id ON public.entitie
 --
 
 CREATE UNIQUE INDEX index_entities_on_uuid ON public.entities USING btree (uuid);
+
+
+--
+-- Name: index_event_store_events_in_streams_on_created_at; Type: INDEX; Schema: public; Owner: -
+--
+
+CREATE INDEX index_event_store_events_in_streams_on_created_at ON public.event_store_events_in_streams USING btree (created_at);
+
+
+--
+-- Name: index_event_store_events_in_streams_on_stream_and_event_id; Type: INDEX; Schema: public; Owner: -
+--
+
+CREATE UNIQUE INDEX index_event_store_events_in_streams_on_stream_and_event_id ON public.event_store_events_in_streams USING btree (stream, event_id);
+
+
+--
+-- Name: index_event_store_events_in_streams_on_stream_and_position; Type: INDEX; Schema: public; Owner: -
+--
+
+CREATE UNIQUE INDEX index_event_store_events_in_streams_on_stream_and_position ON public.event_store_events_in_streams USING btree (stream, "position");
+
+
+--
+-- Name: index_event_store_events_on_created_at; Type: INDEX; Schema: public; Owner: -
+--
+
+CREATE INDEX index_event_store_events_on_created_at ON public.event_store_events USING btree (created_at);
+
+
+--
+-- Name: index_event_store_events_on_event_id; Type: INDEX; Schema: public; Owner: -
+--
+
+CREATE UNIQUE INDEX index_event_store_events_on_event_id ON public.event_store_events USING btree (event_id);
+
+
+--
+-- Name: index_event_store_events_on_event_type; Type: INDEX; Schema: public; Owner: -
+--
+
+CREATE INDEX index_event_store_events_on_event_type ON public.event_store_events USING btree (event_type);
+
+
+--
+-- Name: index_event_store_events_on_valid_at; Type: INDEX; Schema: public; Owner: -
+--
+
+CREATE INDEX index_event_store_events_on_valid_at ON public.event_store_events USING btree (valid_at);
 
 
 --
@@ -1524,6 +1669,7 @@ INSERT INTO "schema_migrations" (version) VALUES
 ('20230222145138'),
 ('20230222161252'),
 ('20230224044112'),
-('20230225034617');
+('20230225034617'),
+('20230225111223');